server/requirements.txt
fastapi
uvicorn
supabase==2.11.0
httpx
pydantic
pydantic-settings
python-dotenv

# Multimodal Context Memory System
PyMuPDF==1.24.0
Pillow==10.3.0
faster-whisper==1.0.1
google-genai>=1.0.0
google-generativeai>=0.5.0
python-multipart==0.0.9
